Start replacing types with stdint types
Replace BYTE with uint8_t

Replace WORD with uint16_t

Replace DWORD with uint32_t

Replace UINT/LONG to uint32_t/long

Replace other types to stdint as well (*/ const * types etc)

+++ libretro/dswin.c
@@ -34,22 +34,22 @@
 int playing = FALSE;
 
 #define PCMBUF_SIZE 2 * 2 * 48000
-BYTE pcmbuffer[PCMBUF_SIZE];
-BYTE *pcmbufp = pcmbuffer;
-BYTE *pbsp = pcmbuffer;
-BYTE *pbrp = pcmbuffer, *pbwp = pcmbuffer;
-BYTE *pbep = &pcmbuffer[PCMBUF_SIZE];
-DWORD ratebase = 22050;
+uint8_t pcmbuffer[PCMBUF_SIZE];
+uint8_t *pcmbufp = pcmbuffer;
+uint8_t *pbsp = pcmbuffer;
+uint8_t *pbrp = pcmbuffer, *pbwp = pcmbuffer;
+uint8_t *pbep = &pcmbuffer[PCMBUF_SIZE];
+uint32_t ratebase = 22050;
 long DSound_PreCounter = 0;
-BYTE rsndbuf[PCMBUF_SIZE];
+uint8_t rsndbuf[PCMBUF_SIZE];
 
 int audio_fd = 1;
 
 void raudio_callback(void *userdata, unsigned char *stream, int len);
 
-int DSound_Init(unsigned long rate, unsigned long buflen)
+int DSound_Init(uint32_t rate, uint32_t buflen)
 {
-	DWORD samples;
+	uint32_t samples;
 
 	if (playing)
 		return FALSE;
@@ -72,8 +72,8 @@ void DSound_Play(void)
 {
 	if (audio_fd >= 0)
 	{
-		ADPCM_SetVolume((BYTE)Config.PCM_VOL);
-		OPM_SetVolume((BYTE)Config.OPM_VOL);
+		ADPCM_SetVolume((uint8_t)Config.PCM_VOL);
+		OPM_SetVolume((uint8_t)Config.OPM_VOL);
 	}
 }
 
@@ -106,7 +106,7 @@ static void sound_send(int length)
 	//Mcry_Update((int16_t *)pcmbufp, length);
 #endif
 
-	pbwp += length * sizeof(WORD) * 2;
+	pbwp += length * sizeof(uint16_t) * 2;
 	if (pbwp >= pbep)
 		pbwp = pbsp + (pbwp - pbep);
 }
@@ -175,7 +175,7 @@ void audio_samples_discard(int discard)
 void raudio_callback(void *userdata, unsigned char *stream, int len)
 {
 	int lena, lenb, datalen, rate;
-	BYTE *buf;
+	uint8_t *buf;
 
 cb_start:
 	if (pbrp <= pbwp)
